multilingual-rag-system/requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1

# PDF Processing
PyMuPDF==1.23.8
easyocr==1.7.0
Pillow==10.1.0

# Embeddings and Models
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0

# Vector Database
faiss-cpu==1.7.4
chromadb==0.4.18

# LLM Models (Free Hugging Face)
huggingface-hub==0.19.4
bitsandbytes==0.41.3

# Reranking
sentence-transformers[rerank]==2.2.2

# UI
streamlit==1.28.1
streamlit-chat==0.1.1

# Utilities
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
python-dotenv==1.0.0

# Language Detection
langdetect==1.0.9

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Evaluation
ragas==0.1.0
datasets==2.15.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
